+++ test/smokes/eslint/expectations.rb
@@ -431,5 +431,5 @@ Smoke.add_test(
 Smoke.add_test(
   "duplicate_lock_files",
   { guid: "test-guid", timestamp: :_, type: "success", issues: [], analyzer: { name: "ESLint", version: "5.16.0" } },
-  { warnings: [{ message: /Two lock files `package-lock.json` and `yarn.lock` are found/, file: nil }] }
+  { warnings: [{ message: /Two lock files `package-lock.json` and `yarn.lock` are found/, file: "yarn.lock" }] }
 )